requirements.txt
#
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License").
# You may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.2.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
audioread=3.0.0=pypi_0
bitarray=2.6.0=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.9.24=ha878542_0
cachetools=4.2.4=pypi_0
certifi=2022.9.24=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
colorama=0.4.5=pypi_0
cudatoolkit=10.2.89=h713d32c_10
cycler=0.11.0=pypi_0
dataclasses=0.6=pypi_0
decorator=5.1.1=pypi_0
distance=0.1.3=pypi_0
fairseq=0.12.2=pypi_0
fastbpe=0.1.0=pypi_0
freetype=2.12.1=hca18f0e_0
future=0.18.2=pypi_0
g2p-en=2.1.0=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.49.1=pypi_0
hydra-core=1.0.7=pypi_0
idna=3.4=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.9.0=pypi_0
inflect=4.1.0=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_2
kiwisolver=1.4.4=pypi_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libdeflate=1.14=h166bdaf_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libnsl=2.0.0=h7f98852_0
libpng=1.6.38=h753d276_0
librosa=0.7.2=pypi_0
libsqlite=3.39.3=h753d276_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.4.0=h55922b4_4
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libzlib=1.2.12=h166bdaf_3
llvm-openmp=14.0.4=he0ac6c6_0
llvmlite=0.31.0=pypi_0
lxml=4.9.1=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.2.2=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mock=4.0.3=pypi_0
ncurses=6.3=h27087fc_1
ninja=1.10.2.4=pypi_0
nltk=3.7=pypi_0
numba=0.48.0=pypi_0
numpy=1.19.0=pypi_0
oauthlib=3.2.1=pypi_0
omegaconf=2.0.6=pypi_0
openjpeg=2.5.0=h7d73246_1
openssl=3.0.5=h166bdaf_2
pillow=9.2.0=py38ha3b2c9c_2
pip=22.2.2=pyhd8ed1ab_0
portalocker=2.5.1=pypi_0
protobuf=3.20.1=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pyparsing=3.0.9=pypi_0
pypinyin=0.39.0=pypi_0
python=3.8.13=ha86cf86_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pyworld=0.2.10=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1.2=h0f457ee_0
regex=2022.9.13=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.3.1=pypi_0
rsa=4.9=pypi_0
sacrebleu=2.2.1=pypi_0
sacremoses=0.0.53=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.0=pypi_0
setuptools=65.4.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.39.3=h4ff8645_0
subword-nmt=0.3.8=pypi_0
tabulate=0.8.10=pypi_0
tbb=2021.6.0=h924138e_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
tgt=1.4.4=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h27826a3_0
torchaudio=0.12.1=pypi_0
torchvision=0.2.2=py_3
tqdm=4.46.1=pypi_0
typing-extensions=4.3.0=pypi_0
typing_extensions=4.4.0=pyha770c72_0
unidecode=1.1.1=pypi_0
urllib3=1.26.12=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
zipp=3.8.1=pypi_0
zstd=1.5.2=h6239696_4
